Allow conditionally disabling various functionality
Most of this is developer-oriented except the first item:
-  If a `ManagerHost` is not available (the cmsd setup is unique to
   the OSDF), then disable running a cmsd.  The cmsd isn't going to
   be used long-term as we switch to prefix-based routing with the
   Director.
-  Provide various boolean config overrides that disable non-critical
   XRootD plugins (Macaroons, VOMS).  Particularly, these can be a pain
   to get installed on a Mac.
-  Auto-generate a CA certificate file.  The `xrootd` process refuses
   to start without at least a CA certificate file.  These are basically
   universally available on Linux but missing on Mac.  Generate a throw-
   away CA file to make XRootD happy.  NOTE: as follow-up, we could
   use this CA to generate the hostcert instead of a self-signed certificate.
   That would unlock end-to-end testing with TLS enabled.

With these changes, I can launch an origin on a Mac laptop on an airplane.
While that's an accomplishment on its own, the real advantage is getting
a step closer toward self-contained integration test environment.

+++ xrootd/launch.go
@@ -150,13 +150,17 @@ func LaunchXrootd(privileged bool, configPath string) (err error) {
 		return err
 	}
 
-	cmsdCtx, cmsdPid, err := launcher.Launch(ctx, "cmsd", configPath)
-	if err != nil {
-		return errors.Wrap(err, "Failed to launch cmsd daemon")
-	}
-	log.Info("Successfully launched cmsd")
-	if err := metrics.SetComponentHealthStatus("cmsd", "ok", ""); err != nil {
-		return err
+	cmsdCtx := context.Background()
+	cmsdPid := -1
+	if viper.GetBool("Origin.UseCmsd") {
+		cmsdCtx, cmsdPid, err = launcher.Launch(ctx, "cmsd", configPath)
+		if err != nil {
+			return errors.Wrap(err, "Failed to launch cmsd daemon")
+		}
+		log.Info("Successfully launched cmsd")
+		if err := metrics.SetComponentHealthStatus("cmsd", "ok", ""); err != nil {
+			return err
+		}
 	}
 
 	sigs := make(chan os.Signal, 1)
@@ -172,8 +176,10 @@ func LaunchXrootd(privileged bool, configPath string) (err error) {
 				if err = syscall.Kill(xrootdPid, sys_sig); err != nil {
 					return errors.Wrap(err, "Failed to forward signal to xrootd process")
 				}
-				if err = syscall.Kill(cmsdPid, sys_sig); err != nil {
-					return errors.Wrap(err, "Failed to forward signal to cmsd process")
+				if cmsdPid != -1 {
+					if err = syscall.Kill(cmsdPid, sys_sig); err != nil {
+						return errors.Wrap(err, "Failed to forward signal to cmsd process")
+					}
 				}
 			} else {
 				panic(errors.New("Unable to convert signal to syscall.Signal"))
@@ -212,7 +218,7 @@ func LaunchXrootd(privileged bool, configPath string) (err error) {
 					return errors.Wrap(err, "Failed to SIGKILL the xrootd process")
 				}
 			}
-			if !cmsdExpiry.IsZero() && time.Now().After(cmsdExpiry) {
+			if cmsdPid != -1 && !cmsdExpiry.IsZero() && time.Now().After(cmsdExpiry) {
 				if err = syscall.Kill(cmsdPid, syscall.SIGKILL); err != nil {
 					return errors.Wrap(err, "Failed to SIGKILL the cmsd process")
 				}
